jangoadmin/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0
django-softdelete==0.9.0
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.9
Markdown==3.2.1
mdown==2019.4.13
mkdir==2019.4.13
mysqlclient==1.4.6
orderdict==2019.9.25
public==2019.4.13
python-decouple==3.3
pytz==2019.3
readme-md==2019.1.28
requests==2.23.0
setupcfg==2019.4.13
sqlparse==0.3.1
urllib3==1.25.8
values==2019.4.13
write==2019.4.13
